Add files via upload
Tic Tac Toe is the game which everybody has played once in their life. This game is coded in C++ Programming language. This is a two player game where the first player can choose its position and place the mark. In the next turn the second player  will get chance to choose the position. This will continue till all the positions are occupied or if it aligns  diagonally, vertically or horizontally.